wydruki_api-master/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
evdev>=1.6.0
mysql-connector-python>=8.0.0
pynput>=1.7.0
tk
python-dotenv>=0.19.0
